Before Implementing inheritance

+++ Controllers/DepartmentsController.cs
@@ -175,7 +175,7 @@ namespace ContosoUniversity.Models
         }
 
         // GET: Departments/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id ,bool? concurrencyError)
         {
             if (id == null)
             {
@@ -184,24 +184,48 @@ namespace ContosoUniversity.Models
 
             var department = await _context.Departments
                 .Include(d => d.Administrator)
+                .AsNoTracking()
                 .SingleOrDefaultAsync(m => m.DepartmentID == id);
             if (department == null)
             {
+                if(concurrencyError.GetValueOrDefault())
+                {
+                    return RedirectToAction(nameof(Index));
+                }
                 return NotFound();
+            }
+            if(concurrencyError.GetValueOrDefault())
+            {
+                ViewData["ConcurrencyErrorMessage"] = "The record you attemped to delete"
+                + "was modified by another user after you got the original values. "
+                + "Teh delete operation was canceled and the the current values in the "
+                + "database have been displayed. If you still want to delete this "
+                + "record, click the Delete button again. Otherwise"
+                + "click the Back to List hyperlink.";
             }
 
             return View(department);
         }
 
         // POST: Departments/Delete/5
-        [HttpPost, ActionName("Delete")]
+        [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> Delete(Department department)
         {
-            var department = await _context.Departments.SingleOrDefaultAsync(m => m.DepartmentID == id);
-            _context.Departments.Remove(department);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+         try 
+         {
+             if (await _context.Departments.AnyAsync(m =>m.DepartmentID==department.DepartmentID))
+             {
+                 _context.Departments.Remove(department);
+                 await _context.SaveChangesAsync();
+             }
+             return RedirectToAction(nameof(Index));
+         }
+         catch(DbUpdateConcurrencyException /* ex */)
+         {
+             // Log the error
+             return RedirectToAction(nameof(Delete),new { concurrencyError = true, id = department.DepartmentID});
+         }
         }
 
         private bool DepartmentExists(int id)
